requirements.txt
tqdm==4.62.0
torch==1.9.0
opencv-python==4.1.2.30
albumentations==0.1.12
flask
numpy
google-cloud-storage
gcsfs
pickle5
pandas
sklearn
kfp